pyproject.toml
[project]
name = "ai-agents-crewai-personalisation"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = {extras = ["tools","agentops"], version = "^0.85.0"}
python-dotenv = "^1.0.1"
langchain-community = "^0.3.19"
langchain-ollama = "^0.2.3"



[tool.poetry]
packages = [{include = "ai_agents_crewai_personalisation", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
run-agents = "ai_agents_crewai_personalisation.main:run"